New api versión

See merge request raupulus/api-fryntiz!9

+++ database/migrations/2022_01_30_232823_create_hardware_power_generators_today_table.php
@@ -66,7 +66,7 @@ class CreateHardwarePowerGeneratorsTodayTable extends Migration
                 ->nullable()
                 ->default(0)
                 ->comment('Porcentaje de batería máximo (%)');
-            $table->decimal('amperage', 10, 2)
+            $table->double('amperage')
                 ->nullable()
                 ->default(0)
                 ->comment('Amperaje total (A)');
